[PRFC] Export kubernetes types and apirefs on path
