+++ plugins/kubernetes-react/src/hooks/auth.ts
@@ -16,8 +16,7 @@
 
 import { Entity } from '@backstage/catalog-model';
 import { KubernetesRequestBody } from '@backstage/plugin-kubernetes-common';
-import { KubernetesAuthProvidersApi } from '../kubernetes-auth-provider';
-import { KubernetesApi } from '../api/types';
+import { KubernetesApi, KubernetesAuthProvidersApi } from '../api/types';
 
 export const generateAuth = async (
   entity: Entity,
