+++ plugins/kubernetes-react/src/hooks/useKubernetesObjects.ts
@@ -21,8 +21,7 @@ import { ObjectsByEntityResponse } from '@backstage/plugin-kubernetes-common';
 import { useApi } from '@backstage/core-plugin-api';
 import { generateAuth } from './auth';
 import useAsyncRetry from 'react-use/lib/useAsyncRetry';
-import { kubernetesAuthProvidersApiRef } from '../kubernetes-auth-provider';
-import { kubernetesApiRef } from '../api/types';
+import { kubernetesApiRef, kubernetesAuthProvidersApiRef } from '../api/types';
 
 /**
  *
