+++ plugins/kubernetes/src/api/KubernetesProxyClient.test.ts
