+++ plugins/kubernetes/src/api/formatters/AksClusterLinksFormatter.test.ts
