+++ plugins/kubernetes/src/api/formatters/OpenshiftClusterLinksFormatter.test.ts
