+++ plugins/kubernetes/src/api/formatters/RancherClusterLinksFormatter.test.ts
