+++ plugins/kubernetes/src/api/formatters/StandardClusterLinksFormatter.test.ts
