+++ plugins/kubernetes/src/api/formatters/StandardClusterLinksFormatter.ts
@@ -16,7 +16,7 @@
 import {
   ClusterLinksFormatter,
   ClusterLinksFormatterOptions,
-} from '../../types';
+} from '@backstage/plugin-kubernetes-react/api';
 
 const kindMappings: Record<string, string> = {
   deployment: 'deployment',
