+++ plugins/kubernetes/src/api/kubernetes-auth-provider/AksKubernetesAuthProvider.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 The Backstage Authors
+ * Copyright 2023 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,26 +13,28 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-import { KubernetesAuthProvider } from './types';
+import { OAuthApi } from '@backstage/core-plugin-api';
 import { KubernetesRequestBody } from '@backstage/plugin-kubernetes-common';
+import { KubernetesAuthProvider } from '@backstage/plugin-kubernetes-react/api';
 
-/**
- * No-op KubernetesAuthProvider, authorization will be handled in the kubernetes-backend plugin
- *
- * @public
- */
-export class ServerSideKubernetesAuthProvider
-  implements KubernetesAuthProvider
-{
+/** @public */
+export class AksKubernetesAuthProvider implements KubernetesAuthProvider {
+  constructor(private readonly microsoftAuthApi: OAuthApi) {}
+
   async decorateRequestBodyForAuth(
     requestBody: KubernetesRequestBody,
   ): Promise<KubernetesRequestBody> {
-    // No-op, auth will be taken care of on the server-side
-    return requestBody;
+    return {
+      ...requestBody,
+      auth: { ...requestBody.auth, aks: (await this.getCredentials()).token },
+    };
   }
 
-  async getCredentials(): Promise<{}> {
-    return {};
+  async getCredentials(): Promise<{ token?: string }> {
+    return {
+      token: await this.microsoftAuthApi.getAccessToken(
+        '6dae42f8-4368-4678-94ff-3960e28e3630/user.read',
+      ),
+    };
   }
 }
