+++ plugins/kubernetes/src/components/KubernetesContent.tsx
