+++ plugins/kubernetes/src/routes.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 The Backstage Authors
+ * Copyright 2024 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,10 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+import { createRouteRef } from '@backstage/core-plugin-api';
 
-export * from './types';
-export * from './KubernetesAuthProviders';
-export * from './GoogleKubernetesAuthProvider';
-export * from './ServerSideAuthProvider';
-export * from './OidcKubernetesAuthProvider';
-export * from './AksKubernetesAuthProvider';
+export const rootCatalogKubernetesRouteRef = createRouteRef({
+  id: 'kubernetes',
+});
